auth 수정
auth service 수정에 따른 사용처 호출 수정
( currentUser 에서 userField 를 가지고 있도록 함 )
