+++ JUDA/Resource/Components/DrinkListCell.swift
@@ -98,7 +98,7 @@ struct DrinkListCell: View {
                     // 디바운서 콜
                     debouncer.call {
                         authService.addOrRemoveToLikedDrinks(isLiked: isLiked, drink.drinkID)
-                        authService.userLikedDrinksUpdate()
+                        authService.userLikedListUpdate(type: .drinks)
                     }
                 } label: {
                     Image(systemName: isLiked ? "heart.fill" : "heart")
@@ -110,7 +110,9 @@ struct DrinkListCell: View {
         .padding(.horizontal, 20)
         .frame(height: 130)
 		.task {
-			self.isLiked = authService.likedDrinks.contains(where: { $0 == drink.drinkID })
+            if let user = authService.currentUser {
+                self.isLiked = user.likedDrinks.contains { $0 == drink.drinkID }
+            }
 		}
     }
 }
