+++ JUDA/View/Posts/PostDetail/PostInfo.swift
@@ -94,8 +94,10 @@ struct PostInfo: View {
         .padding(.horizontal, 20)
         .padding(.vertical, 5)
 		.task {
-			self.isLike = authService.likedPosts.contains(where: { $0 == post.postField.postID })
-			self.likeCount = post.postField.likedCount
+            if let user = authService.currentUser {
+                self.isLike = user.likedPosts.contains { $0 == post.postField.postID }
+                self.likeCount = post.postField.likedCount
+            }
 		}
     }
 	
@@ -109,8 +111,8 @@ struct PostInfo: View {
 		Task {
 			if isLike {
 				likeCount -= 1
-				authService.likedPosts.removeAll(where: { $0 == postID })
-				authService.userLikedPostsUpdate()
+                authService.currentUser?.likedPosts.removeAll { $0 == postID }
+                authService.userLikedListUpdate(type: .posts)
 				
 				if let index = searchPostsViewModel.posts.firstIndex(where: { $0.postField.postID == postID }) {
 					searchPostsViewModel.posts[index].postField.likedCount -= 1
@@ -146,8 +148,8 @@ struct PostInfo: View {
 				await postsViewModel.postLikedUpdate(likeType: .minus, postID: postID, userID: post.userField.userID ?? "")
 			} else {
 				likeCount += 1
-				authService.likedPosts.append(postID)
-				authService.userLikedPostsUpdate()
+                authService.currentUser?.likedPosts.append(postID)
+                authService.userLikedListUpdate(type: .posts)
 				
 				if let index = searchPostsViewModel.posts.firstIndex(where: { $0.postField.postID == postID }) {
 					searchPostsViewModel.posts[index].postField.likedCount += 1
